fix: buttonPrimary56 컴포넌트 수정
