+++ src/components/chooseBackground/chooseBackground.jsx
@@ -1,3 +1,7 @@
+import { useNavigate } from 'react-router-dom';
+import { postApiRecipient } from '../../apis/apiRecipient';
+import { useSetRecoilState } from 'recoil';
+import { userId } from '../../store/recoil/apiData';
 import ToggleButton from '../toggleButton/toggleButton';
 import ButtonPrimary56 from '../button/buttonPrimary/buttonPrimary56/buttonPrimary56';
 import './chooseBackground.css';
@@ -9,6 +13,28 @@ function ChooseBackground({
   setBackgroundColor,
   setBackgroundImg,
 }) {
+  const navigate = useNavigate();
+  const setSelectId = useSetRecoilState(userId);
+
+  const userData = {
+    name: !!userName === false ? null : userName,
+    color: !!backgroundColor === false ? 'beige' : backgroundColor,
+    img: !!backgroundImg === false ? null : backgroundImg,
+  };
+
+  const handlePostUserInfo = () => {
+    if (userData.name === null) {
+      //데이터 보내면 안됨
+    } else {
+      postApiRecipient(userData.name, userData.color, userData.backgroundImg).then((response) => {
+        const data = response;
+        setSelectId(data.id);
+        // post/{id} 페이지 이동
+        navigate(`/post/${data.id}`);
+      });
+    }
+  };
+
   return (
     <div className="chooseBackground">
       <div className="chooseBackgroundText">
@@ -16,12 +42,7 @@ function ChooseBackground({
         <p>컬러를 선택하거나, 이미지를 선택할 수 있습니다.</p>
       </div>
       <ToggleButton setBackgroundColor={setBackgroundColor} setBackgroundImg={setBackgroundImg} />
-      <ButtonPrimary56
-        buttonName="생성하기"
-        userName={userName}
-        backgroundColor={backgroundColor}
-        backgroundImg={backgroundImg}
-      />
+      <ButtonPrimary56 buttonName="생성하기" handleApi={handlePostUserInfo} />
     </div>
   );
 }
